refactor: Move `_ImplDescriptor` to `_utils`
