+++ narwhals/_utils.py
@@ -74,16 +74,37 @@ if TYPE_CHECKING:
         NativeSeriesT_co,
     )
     from narwhals._compliant.typing import EvalNames, NativeLazyFrameT
-    from narwhals._namespace import Namespace
+    from narwhals._namespace import (
+        Namespace,
+        _CuDFDataFrame,
+        _ModinDataFrame,
+        _NativeDask,
+        _NativeDuckDB,
+        _NativeIbis,
+        _NativePandasLikeDataFrame,
+        _NativePandasLikeSeries,
+        _NativeSQLFrame,
+    )
     from narwhals._translate import ArrowStreamExportable, IntoArrowTable, ToNarwhalsT_co
     from narwhals._typing import (
         Backend,
         IntoBackend,
+        _ArrowImpl,
+        _CudfImpl,
+        _DaskImpl,
         _DataFrameLazyImpl,
+        _DuckDBImpl,
         _EagerAllowedImpl,
+        _IbisImpl,
+        _LazyAllowedImpl,
         _LazyFrameCollectImpl,
+        _ModinImpl,
+        _PandasImpl,
+        _PandasLikeImpl,
+        _PolarsImpl,
+        _SQLFrameImpl,
     )
-    from narwhals.dataframe import DataFrame, LazyFrame
+    from narwhals.dataframe import BaseFrame, DataFrame, LazyFrame
     from narwhals.dtypes import DType
     from narwhals.series import Series
     from narwhals.typing import (
@@ -2032,3 +2053,67 @@ def deep_attrgetter(attr: str, *nested: str) -> attrgetter[Any]:
 def deep_getattr(obj: Any, name_1: str, *nested: str) -> Any:
     """Perform a nested attribute lookup on `obj`."""
     return deep_attrgetter(name_1, *nested)(obj)
+
+
+class _ImplDescriptor:
+    def __set_name__(self, owner: type[Any], name: str) -> None:
+        self.__name__: str = name
+
+    @overload
+    def __get__(
+        self,
+        instance: DataFrame[pl.DataFrame] | LazyFrame[pl.LazyFrame] | Series[pl.Series],
+        owner: Any,
+    ) -> _PolarsImpl: ...
+    @overload
+    def __get__(
+        self, instance: BaseFrame[pd.DataFrame] | Series[pd.DataFrame], owner: Any
+    ) -> _PandasImpl: ...
+    @overload
+    def __get__(self, instance: BaseFrame[_ModinDataFrame], owner: Any) -> _ModinImpl: ...
+
+    @overload  # oof, looks like these two need their names aligned 😅
+    def __get__(self, instance: BaseFrame[_CuDFDataFrame], owner: Any) -> _CudfImpl: ...
+    @overload
+    def __get__(
+        self,
+        instance: BaseFrame[_NativePandasLikeDataFrame] | Series[_NativePandasLikeSeries],
+        owner: Any,
+    ) -> _PandasLikeImpl: ...
+    @overload
+    def __get__(
+        self, instance: BaseFrame[pa.Table] | Series[pa.ChunkedArray[Any]], owner: Any
+    ) -> _ArrowImpl: ...
+    @overload
+    def __get__(
+        self,
+        instance: BaseFrame[pl.DataFrame | pd.DataFrame | pa.Table]
+        | Series[pl.Series | pd.Series[Any] | pa.ChunkedArray[Any]],
+        owner: Any,
+    ) -> _PolarsImpl | _PandasImpl | _ArrowImpl: ...
+    @overload
+    def __get__(self, instance: LazyFrame[_NativeDuckDB], owner: Any) -> _DuckDBImpl: ...
+    @overload
+    def __get__(
+        self, instance: LazyFrame[_NativeSQLFrame], owner: Any
+    ) -> _SQLFrameImpl: ...
+    @overload
+    def __get__(self, instance: LazyFrame[_NativeDask], owner: Any) -> _DaskImpl: ...
+    @overload
+    def __get__(self, instance: LazyFrame[_NativeIbis], owner: Any) -> _IbisImpl: ...
+    @overload
+    def __get__(self, instance: None, owner: Any) -> Self: ...
+    @overload
+    def __get__(
+        self, instance: DataFrame[Any] | Series[Any], owner: Any
+    ) -> _EagerAllowedImpl: ...
+    @overload
+    def __get__(self, instance: LazyFrame[Any], owner: Any) -> _LazyAllowedImpl: ...
+    def __get__(
+        self,
+        instance: DataFrame[Any] | LazyFrame[Any] | BaseFrame[Any] | Series[Any] | None,
+        owner: Any,
+    ) -> Any:
+        if instance is None:  # pragma: no cover
+            return self
+        return instance._compliant._implementation
